moved actions and svg to its own folder, added interface to delete form btn
