+++ app/components/svg/laptop-svg.tsx
