+++ app/components/svg/logo-svg.tsx
